chore: move files to more appropriate places after code review
