+++ src/Ushahidi/Modules/V5/Http/Controllers/CountryCodeController.php
@@ -7,8 +7,8 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Gate;
-use Ushahidi\Modules\V5\Actions\CountryCode\Query\FetchCountryCodeByIdQuery;
-use Ushahidi\Modules\V5\Actions\CountryCode\Query\FetchCountryCodeQuery;
+use Ushahidi\Modules\V5\Actions\CountryCode\Queries\FetchCountryCodeByIdQuery;
+use Ushahidi\Modules\V5\Actions\CountryCode\Queries\FetchCountryCodeQuery;
 use Ushahidi\Modules\V5\Http\Resources\CountryCodeCollection;
 use Ushahidi\Modules\V5\Http\Resources\CountryCodeResource;
 use Ushahidi\Modules\V5\Models\CountryCode;
